server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.13.1
PyJWT==2.10.1
authlib==1.6.9
pymongo==4.16.0
redis==6.4.0
httpx==0.28.1
dashscope>=1.20.0
dashvector>=1.0.0
pytest==7.3.1